% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_builder.R
\name{glb_summary}
\alias{glb_summary}
\title{General summary table}
\usage{
glb_summary(data, y)
}
\arguments{
\item{data}{data. Must be in the tidy format}

\item{y}{response variable}
}
\description{
General summary table
}
\author{
Omar Benites
}

