% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_builder.R
\name{grp_summary}
\alias{grp_summary}
\title{Summary table by group}
\usage{
grp_summary(data, group, y)
}
\arguments{
\item{data}{data. Must be in the tidy format}

\item{group}{group by}

\item{y}{response variable
# @param test statistical test that return an specific table. It works with \code{Median}, \code{Jonckheere},...}
}
\description{
Summary table by group
}
\author{
Omar Benites
}

