% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converters.R
\name{rename_tables}
\alias{rename_tables}
\title{Rename tables}
\usage{
rename_tables(dt, new_names)
}
\arguments{
\item{dt}{data table}

\item{new_names}{new column names}
}
\description{
Rename tables
}
\author{
Omar Benites
}

