% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_builder.R
\name{test_analysis}
\alias{test_analysis}
\title{Return tables (data frames) derived from statistical analysis}
\usage{
test_analysis(x, y, hyp, param, jud, test = "friedman", sg = 0.05,
  comparison = FALSE)
}
\arguments{
\item{x}{independent variable}

\item{y}{dependent variable}

\item{hyp}{hypothethical testing}

\item{param}{statistical parameter such as \code{mu}, \code{median}, etc.}

\item{jud}{judges.}

\item{test}{statistical test. Select \code{median}, \code{durbin}
\code{jonckheere}, \code{kruskal}, \code{friedman}, \code{manwithney} and \code{wilcoxon}}

\item{sg}{Significant level. By defualt \code{alpha=0.05}}

\item{comparison}{Wheter calculate comparison between samples means by treatment.}
}
\description{
Statistical tables from non-parametric analysis
}
\author{
Omar Benites
}

