% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_builder.R
\name{test_summary}
\alias{test_summary}
\title{Summary table}
\usage{
test_summary(data, group, y, test = "jonk")
}
\arguments{
\item{data}{data. Must be in the tidy format}

\item{group}{group by}

\item{y}{response variable}

\item{test}{statistical test that return an specific table. It works with \code{Median}, \code{Jonckheere},...}
}
\description{
Summary table
}
\author{
Omar Benites
}

