% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converters.R
\name{agr2df}
\alias{agr2df}
\title{Converter for agricolae's table to data.frame}
\usage{
agr2df(adt)
}
\arguments{
\item{adt}{agricolae's data table}
}
\description{
Converter for agricolae's table to data.frame
}
\author{
Omar Benites
}

